fix(mcp): address PR review — intent tool name, version comment, LICENSE, test dedup
- intent: resolve the missing-capability probe through
  resolve_missing_capability_tool_name() instead of the hardcoded
  "get_more_tools", so a custom missing_capability_tool_name no longer
  misclassifies its calls as context_parameter intent.
- version: fix the comment — __version__ IS stamped as sdk_version on
  every captured event (capture.py), it is not informational-only.
- LICENSE: add the MCPCat MIT attribution block (code derived from
  MCPCat/mcpcat-typescript-sdk), mirroring the per-file headers and the
  posthog-js LICENSE.
- tests: extract the duplicated FakeClient / flush / events helpers into
  posthog/test/mcp/_helpers.py (OnceAndOnlyOnce) and import them.

Generated-By: PostHog Code
Task-Id: b21bc954-5de3-4512-a0d5-6bec2371f782
