+++ posthog/mcp/version.py
@@ -2,6 +2,6 @@
 # Copyright (c) 2025 MCPcat
 # Licensed under the MIT License: https://github.com/MCPCat/mcpcat-typescript-sdk/blob/main/LICENSE
 
-# Version of the PostHog MCP analytics SDK surface. Informational only — it is
-# not stamped onto captured events. The package itself ships inside posthog.
+# Version of the PostHog MCP analytics SDK surface. Stamped as ``sdk_version``
+# on every captured event. The package itself ships inside posthog.
 __version__ = "0.1.0"
