+++ posthog/test/mcp/test_lowlevel.py
@@ -1,30 +1,15 @@
 """End-to-end tests for the low-level mcp.server.Server adapter (Milestone 3)."""
 
-import asyncio
 
 import mcp.types as mcp_types
 from mcp.server.lowlevel import Server
 
 from posthog.mcp import instrument
-
-
-class FakeClient:
-    def __init__(self):
-        self.events = []
-
-    def capture(
-        self,
-        event,
-        distinct_id=None,
-        properties=None,
-        timestamp=None,
-        uuid=None,
-        **kwargs,
-    ):
-        self.events.append(
-            {"event": event, "distinct_id": distinct_id, "properties": properties or {}}
-        )
-        return None
+from posthog.test.mcp._helpers import (
+    FakeClient,
+    events_named as _events,
+    flush_background as _flush,
+)
 
 
 def make_server():
@@ -51,21 +36,6 @@ def make_server():
         raise ValueError("boom")
 
     return server
-
-
-async def _flush():
-    import posthog.mcp.instrumentation as instr
-
-    for _ in range(10):
-        await asyncio.sleep(0)
-        pending = [t for t in list(instr._BACKGROUND_TASKS) if not t.done()]
-        if pending:
-            await asyncio.gather(*pending, return_exceptions=True)
-    await asyncio.sleep(0)
-
-
-def _events(client, name):
-    return [e for e in client.events if e["event"] == name]
 
 
 def _call_request(name, arguments):
